Logical backup to include all database --single-transaction
Using --single-transaction to ensure minimal impact and adding
--events/--routines to ensure databackup is complete.

Added --master-data=2/--dump-slave=2 with gtid so the data is there
to enable replication recovery.

Added DumpOpts for a user changeable configuration at which the
defaults are replaced fully.

Use 'mariadb' instead of 'mysql' for the restore along with
mariadb-dump instead of mysqldump.

Closes: #65