refactor: Changed the Replication package to use the new AuthController
This is done instead of the Replication package having it's own logic. Notable the replication package is using the
new SqlStrategy, as the creation needs to happen with the `headless` service, where normally it would use the cluster svc.

The problem with using the cluter one is that the replication pods will not be ready, hence user creation will not happen as the svc will
not be reachable.