feat: Migration script for v25.10.0
Does an in-file replacement of properties if they are present only.
Creates the changes in a new file rather than modifing the original one.
Will show a nice summary of what was done in the end if `diff` tool exists.

Also extracts the `status` field, but to be fair I don't undertand why this is needed, will leave it for the PR review