test: Removed PVC finalizer removal after a timeout
There was an odd behavior that was observed initially when testing. The
PVC would get stuck in Terminating state and would need to be removed.
It is no longer happening (after `make cluster-delete` and `make cluster`),
so I am inclined to believe that it was happening due to a "dirty" environment,
hard to say.